Allow forcing MacVim to run g:rspec_command within vim
This allows one to continue using vim-dispatch as the command runner. Before this change, vim-rspec assumed that
just because we're running MacVim, we should always shell the command out to <plugin>/bin/<g:rspec_runner>, but
dispatch takes care of running the command for us, with the added benefit of pulling the output back into a
quickfix window.